feat: Update OpenRouter provider to support embeddings and adjust related tests
